namkim : add : chat-events dto : update : chatGateway
