+++ src/backend/package.json
@@ -27,12 +27,13 @@
     "@nestjs/config": "^3.0.0",
     "@nestjs/core": "^10.0.0",
     "@nestjs/jwt": "^10.1.0",
+    "@nestjs/mapped-types": "^2.0.2",
     "@nestjs/passport": "^10.0.0",
     "@nestjs/platform-express": "^10.0.0",
     "@nestjs/platform-socket.io": "^10.1.3",
     "@nestjs/serve-static": "^4.0.0",
     "@nestjs/websockets": "^10.1.3",
-    "@prisma/client": "^5.1.1",
+    "@prisma/client": "^5.2.0",
     "bcrypt": "^5.1.1",
     "class-transformer": "^0.5.1",
     "class-validator": "^0.14.0",
@@ -61,7 +62,7 @@
     "eslint-plugin-prettier": "^4.2.1",
     "jest": "^29.5.0",
     "prettier": "^2.8.8",
-    "prisma": "^5.1.1",
+    "prisma": "^5.2.0",
     "source-map-support": "^0.5.21",
     "supertest": "^6.3.3",
     "ts-jest": "^29.1.0",
