--- src/backend/src/socket.io/chat/chat.gateway2.ts
@@ -1,180 +1,0 @@
-import { Logger } from '@nestjs/common';
-import {
-  ConnectedSocket,
-  OnGatewayConnection,
-  OnGatewayDisconnect,
-  OnGatewayInit,
-  SubscribeMessage,
-  WebSocketGateway,
-  WebSocketServer,
-} from '@nestjs/websockets';
-import { Namespace } from 'socket.io';
-import { ChatService } from './chat.service';
-import { ChatSocket } from './types';
-import { MessageDto, RoomDto, RoomEventDto, TargetDto } from './dto/chat-events.dto';
-
-@WebSocketGateway({
-	namespace: 'chat',
-})
-export class ChatGateway
-  implements OnGatewayConnection, OnGatewayDisconnect, OnGatewayInit
-{
-	@WebSocketServer() io : Namespace;
-	private readonly logger = new Logger(ChatGateway.name);
-	constructor(private chatService: ChatService) {}
-
-	afterInit(){
-		this.chatService.initChatServer();
-	}
-
-	handleConnection(@ConnectedSocket() client: ChatSocket) {
-		//db 랑 연동할 것인지...?
-	    this.logger.log(`Client Connected : ${client.nickname}, ${client.id}`);
-	
-		client.onAny((any : any) => {
-			this.logger.log(`From Client ${client.nickname}, ${client.id}, event : ${any}`);
-		})
-		this.chatService.handleNewConnection(this.io, client);
-	}
-
-	async handleDisconnect(@ConnectedSocket() client: ChatSocket) {
-		this.logger.log(client.nickname + ", " + client.id + ' is leaving');
-		await this.chatService.handleDisconnection(this.io, client);
-	}
-
-
-	@SubscribeMessage('sendChatMessage')
-	handleSendChat(client: ChatSocket, payload : MessageDto){
-		this.chatService.sendChat(
-			this.io, client, payload.to, payload.body
-		);
-	}
-
-	@SubscribeMessage('sendDirectMessage')
-	handleSendDM(client: ChatSocket, payload : MessageDto){
-		this.chatService.fetchDM(
-			this.io, client, payload.to, payload.body
-		);
-	}
-
-	@SubscribeMessage('selectRoom')
-	handleEnterRoom(client: ChatSocket, payload : RoomDto){
-		this.chatService.userJoinRoom(
-			this.io, client, payload.roomname
-		);
-	}
-
-	@SubscribeMessage('sendRoomLeave')
-	handleLeaveRoom(client: ChatSocket, payload : RoomDto){
-		this.chatService.userLeaveRoom(this.io, client, payload.roomname);
-    	this.chatService.userLeaveRoomAct(this.io, client.userId, payload.roomname);
-	}
-
-	@SubscribeMessage('sendRoomPass')
-	handleEnterRoomWithPassword(client: ChatSocket, payload : RoomDto){
-		this.chatService.userJoinRoom(
-			this.io, client, payload.roomname, payload.password
-		);
-	}
-
-	@SubscribeMessage('setRoomPass')
-	handleSetRoomPass(client: ChatSocket, payload : RoomDto){
-		this.chatService.setPassword(
-			this.io, client, payload.roomname, payload.password
-		);
-	}
-
-	@SubscribeMessage('setRoomPrivate')
-	handleSetRoomPrivate(client: ChatSocket, payload : RoomDto){
-		this.chatService.setRoomStatus(
-			this.io, client, payload.roomname, true
-		);
-	}
-	@SubscribeMessage('setRoomPublic')
-	handleSetRoomPulbic(client: ChatSocket, payload : RoomDto){
-		this.chatService.setRoomStatus(
-			this.io, client, payload.roomname, false
-		);
-	}
-
-	@SubscribeMessage('blockUser')
-	handleBlockUser(client: ChatSocket, payload : TargetDto){
-		this.chatService.blockUser(
-			this.io, client, payload.target
-		);
-	}
-
-	@SubscribeMessage('unblockUser')
-	handleUnblockUser(client: ChatSocket, payload : TargetDto){
-		this.chatService.unblockUser(
-			this.io, client, payload.target
-		);
-	}
-	
-	//과연...?
-	@SubscribeMessage('kickUser')
-	handleKickUser(client: ChatSocket, payload : RoomEventDto){
-		this.chatService.unblockUser(
-			this.io, client, payload.roomname, payload.target
-		);
-	}
-	
-	@SubscribeMessage('banUser')
-	handleBanUser(client: ChatSocket, payload : RoomEventDto){
-		this.chatService.unblockUser(
-			this.io, client, payload.roomname, payload.target
-		);
-	}
-	
-	@SubscribeMessage('muteUser')
-	handleMuteUser(client: ChatSocket, payload : RoomEventDto){
-		this.chatService.unblockUser(
-			this.io, client, payload.roomname, payload.target
-		);
-	}
-	
-	@SubscribeMessage('addOperator')
-	handleAddOper(client: ChatSocket, payload : RoomEventDto){
-		this.chatService.addOperator(
-			this.io, client, payload.roomname, payload.target
-		);
-	}
-	
-	@SubscribeMessage('deleteOperator')
-	handleDelOper(client: ChatSocket, payload : RoomEventDto){
-		this.chatService.deleteOperator(
-			this.io, client, payload.roomname, payload.target
-		);
-	}
-
-	@SubscribeMessage('requestAllRoomList')
-	handleReqAllRoomList(client: ChatSocket){
-		const roomInfo = this.chatService.getAllRoomList(client.userId);
-		client.emit('sendRoomList', roomInfo);
-	}
-	
-	@SubscribeMessage('requestMyRoomList')
-	handleReqUserRoomList(client: ChatSocket){
-    	const roomInfo = this.chatService.getUserRoomList(client.userId);
-		client.emit('sendRoomList', roomInfo);
-
-	}
-	
-	@SubscribeMessage('requestAllMembers')
-	handleReqAllMembers(client: ChatSocket){
-		const members = this.chatService.getAllUserInfo();
-    	client.emit('responseAllMembers', members);
-	}
-	
-	@SubscribeMessage('requestRoomMembers')
-	handleReqRoomMembers(client: ChatSocket, payload : RoomDto){
-		const roomMembers = this.chatService.makeRoomUserInfo(payload.roomname);
-		client.emit('sendRoomMembers', roomMembers);
-	}
-	
-	@SubscribeMessage('requestSearchResultRoomList')
-	handleReqQueryRes(client: ChatSocket, payload : TargetDto){
-		const roomInfo = this.chatService.getQueryRoomList(client.userId, payload.target);
-      client.emit('responseRoomQuery', roomInfo);
-	}
-}
