+++ src/backend/src/socket.io/chat/store/store.module.ts
@@ -3,7 +3,6 @@ import { ChatMessageStoreService } from './store.message.service';
 import { ChatUserStoreService } from './store.user.service';
 import { ChatRoomStoreService } from './store.room.service';
 
-//사용이 좀 이상한 것 같지만...!
 @Module({
 	providers : [
 		ChatMessageStoreService,
